# store-gateway: ExpandedPostings shortcut: avoid LabelValues unless necessary