--- a/pkg/storegateway/bucket_index_postings.go
+++ b/pkg/storegateway/bucket_index_postings.go
@@ -29,41 +29,123 @@ type postingGroup struct {
 	removeKeys []labels.Label
 }
 
-func newPostingGroup(addAll bool, addKeys, removeKeys []labels.Label) *postingGroup {
-	return &postingGroup{
+func newPostingGroup(addAll bool, addKeys, removeKeys []labels.Label) postingGroup {
+	return postingGroup{
 		addAll:     addAll,
 		addKeys:    addKeys,
 		removeKeys: removeKeys,
 	}
 }
 
-// NOTE: Derived from tsdb.postingsForMatcher. index.Merge is equivalent to map duplication.
-func toPostingGroup(lvr indexheader.Reader, m *labels.Matcher) (*postingGroup, error) {
+func (g postingGroup) filterNonExistentKeys(preader postingsOffsetReader) (_ postingGroup, retErr error) {
+	filterKeys := func(keys []labels.Label) []labels.Label {
+		existingKeys := 0
+		for i, l := range keys {
+			if _, err := preader.PostingsOffset(l.Name, l.Value); errors.Is(err, indexheader.NotFoundRangeErr) {
+				// This label name and value doesn't exist in this block, so there are 0 postings we can match.
+				// Set it to an empty value, so we can filter it out later.
+				keys[i] = labels.Label{}
+				// Try with the rest of the set matchers, maybe they can match some series.
+				continue
+			} else if err != nil {
+				retErr = err
+				return nil
+			}
+			existingKeys++
+		}
+		if existingKeys == len(keys) {
+			return keys
+		}
+		if existingKeys == 0 {
+			return nil
+		}
+
+		filtered := make([]labels.Label, 0, existingKeys)
+		for _, k := range keys {
+			if k == (labels.Label{}) {
+				continue
+			}
+			filtered = append(filtered, k)
+		}
+		return filtered
+	}
+
+	return postingGroup{
+		addAll:     g.addAll,
+		addKeys:    filterKeys(g.addKeys),
+		removeKeys: filterKeys(g.removeKeys),
+	}, retErr
+}
+
+type lazyPostingGroup struct {
+	isRemoveGroup bool
+	labelName     string
+	matcher       func(string) bool
+	addAll        bool
+}
+
+func newLazyPostingGroup(addAll bool, isRemoveGroup bool, labelName string, matcher func(string) bool) lazyPostingGroup {
+	return lazyPostingGroup{
+		isRemoveGroup: isRemoveGroup,
+		labelName:     labelName,
+		matcher:       matcher,
+		addAll:        addAll,
+	}
+}
+
+// toPostingGroup guarantees that all the keys in the returned postingGroup exist in the index
+func (g lazyPostingGroup) toPostingGroup(lvr labelValuesReader) (postingGroup, error) {
+	vals, err := lvr.LabelValues(g.labelName, g.matcher)
+	if err != nil {
+		return postingGroup{}, err
+	}
+	keys := make([]labels.Label, len(vals))
+	for i := range vals {
+		keys[i] = labels.Label{Name: g.labelName, Value: vals[i]}
+	}
+
+	if g.isRemoveGroup {
+		return postingGroup{
+			addAll:     g.addAll,
+			removeKeys: keys,
+		}, nil
+	}
+	return postingGroup{
+		addAll:  g.addAll,
+		addKeys: keys,
+	}, nil
+}
+
+type postingsOffsetReader interface {
+	PostingsOffset(name string, value string) (index.Range, error)
+}
+
+type labelValuesReader interface {
+	LabelValues(name string, filter func(string) bool) ([]string, error)
+}
+
+// toPostingGroup returns either a postingGroup or a lazyPostingGroup. It returns a postingGroup when it can
+// derive all the possible values that would satisfy the matcher without looking up the index
+// (e.g. pod="compactor=1"` or pod=~"compactor-(0|1|2|3)")
+// If the matcher needs to use the index, then toPostingGroup returns a lazyPostingGroup (e.g. pod=~"compactor.*").
+// NOTE: Derived from tsdb.postingsForMatcher
+func toPostingGroup(m *labels.Matcher) (_ postingGroup, _ lazyPostingGroup, isLazy bool) {
 	if setMatches := m.SetMatches(); len(setMatches) > 0 && (m.Type == labels.MatchRegexp || m.Type == labels.MatchNotRegexp) {
 		keys := make([]labels.Label, 0, len(setMatches))
 		for _, val := range setMatches {
-			if _, err := lvr.PostingsOffset(m.Name, val); errors.Is(err, indexheader.NotFoundRangeErr) {
-				// This label name and value doesn't exist in this block, so there are 0 postings we can match.
-				// Try with the rest of the set matchers, maybe they can match some series.
-				continue
-			}
 			keys = append(keys, labels.Label{Name: m.Name, Value: val})
 		}
 		if m.Type == labels.MatchNotRegexp {
-			return newPostingGroup(true, nil, keys), nil
+			return newPostingGroup(true, nil, keys), lazyPostingGroup{}, false
 		}
-		return newPostingGroup(false, keys, nil), nil
+		return newPostingGroup(false, keys, nil), lazyPostingGroup{}, false
 	}
 
 	if m.Value != "" {
 		// Fast-path for equal matching.
 		// Works for every case except for `foo=""`, which is a special case, see below.
 		if m.Type == labels.MatchEqual {
-			if _, err := lvr.PostingsOffset(m.Name, m.Value); errors.Is(err, indexheader.NotFoundRangeErr) {
-				// This label name or value doesn't exist in this block, so there are 0 postings we can match.
-				return newPostingGroup(false, nil, nil), nil
-			}
-			return newPostingGroup(false, []labels.Label{{Name: m.Name, Value: m.Value}}, nil), nil
+			return newPostingGroup(false, []labels.Label{{Name: m.Name, Value: m.Value}}, nil), lazyPostingGroup{}, false
 		}
 
 		// If matcher is `label!="foo"`, we select an empty label value too,
@@ -71,12 +153,7 @@ func toPostingGroup(lvr indexheader.Reader, m *labels.Matcher) (*postingGroup, e
 		// So this matcher selects all series in the storage,
 		// except for the ones that do have `label="foo"`
 		if m.Type == labels.MatchNotEqual {
-			if _, err := lvr.PostingsOffset(m.Name, m.Value); errors.Is(err, indexheader.NotFoundRangeErr) {
-				// This label name or value doesn't exist in this block, so we should include all postings
-				// because all series match m.
-				return newPostingGroup(true, nil, nil), nil
-			}
-			return newPostingGroup(true, nil, []labels.Label{{Name: m.Name, Value: m.Value}}), nil
+			return newPostingGroup(true, nil, []labels.Label{{Name: m.Name, Value: m.Value}}), lazyPostingGroup{}, false
 		}
 	}
 
@@ -87,22 +164,12 @@ func toPostingGroup(lvr indexheader.Reader, m *labels.Matcher) (*postingGroup, e
 	// have the label name set too. See: https://github.com/prometheus/prometheus/issues/3575
 	// and https://github.com/prometheus/prometheus/pull/3578#issuecomment-351653555.
 	if m.Matches("") {
-		vals, err := lvr.LabelValues(m.Name, not(m.Matches))
-		toRemove := make([]labels.Label, len(vals))
-		for i := range vals {
-			toRemove[i] = labels.Label{Name: m.Name, Value: vals[i]}
-		}
-		return newPostingGroup(true, nil, toRemove), err
+		return postingGroup{}, newLazyPostingGroup(true, true, m.Name, not(m.Matches)), true
 	}
 
 	// Our matcher does not match the empty value, so we just need the postings that correspond
 	// to label values matched by the matcher.
-	vals, err := lvr.LabelValues(m.Name, m.Matches)
-	toAdd := make([]labels.Label, len(vals))
-	for i := range vals {
-		toAdd[i] = labels.Label{Name: m.Name, Value: vals[i]}
-	}
-	return newPostingGroup(false, toAdd, nil), err
+	return postingGroup{}, newLazyPostingGroup(false, false, m.Name, m.Matches), true
 }
 
 func not(filter func(string) bool) func(string) bool {