… pkg/storegateway/bucket_index_reader.go