This is a proposed implementation of [this comment](https://github.com/grafana/mimir/pull/3848#issuecomment-1371179255) from #3848