@dimitarvdimitrov 2023-01-06
this is `BenchmarkBucketIndexReader_ExpandedPostings` between the tip of https://github.com/grafana/mimir/pull/3848 and this branch


<details>
<summary>BenchmarkBucketIndexReader_ExpandedPostings</summary>

```
name                                                                        old time/op    new time/op    delta
BucketIndexReader_ExpandedPostings/n="X"-10                                    886ns ± 1%    1335ns ±28%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="X",j="foo"-10                            962ns ± 1%    1151ns ± 4%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="X",j!="foo"-10                           957ns ± 1%    1111ns ± 1%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/j=~"XXX|YYY"-10                             954ns ± 1%    1009ns ± 1%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/j=~"X.+"-10                                1.24µs ± 0%    1.34µs ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"X|Y|Z"-10                              1.29µs ± 1%    1.34µs ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1"-10                                   4.81ms ± 0%    4.91ms ± 2%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",j="foo"-10                           45.6ms ± 0%    48.9ms ± 7%   ~     (p=0.200 n=3+3)
BucketIndexReader_ExpandedPostings/j="foo",n="1"-10                           45.7ms ± 0%    46.1ms ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",j!="foo"-10                          36.5ms ± 2%    36.1ms ± 1%   ~     (p=0.700 n=3+3)
BucketIndexReader_ExpandedPostings/i=~".*"-10                                 86.6ms ± 1%    87.1ms ± 1%   ~     (p=0.400 n=3+3)
BucketIndexReader_ExpandedPostings/i=~".+"-10                                  266ms ± 0%     268ms ± 1%   ~     (p=0.400 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"^.+$",j=~"X.+"-10                      5.00ms ±11%    2.99ms ± 1%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~""-10                                    409ms ± 3%     397ms ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i!=""-10                                    268ms ± 1%     267ms ± 1%   ~     (p=0.700 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",j="foo"-10                   48.7ms ± 1%    47.6ms ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="X",i=~"^.+$",j="foo"-10                 1.03µs ± 1%    1.31µs ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",i!="2",j="foo"-10            56.8ms ± 1%    56.4ms ± 0%   ~     (p=0.400 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i!=""-10                              144ms ± 0%     145ms ± 0%   ~     (p=0.700 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i!="",j="foo"-10                      173ms ± 1%     167ms ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i!="",j=~"X.+"-10                    3.73ms ± 1%    2.32ms ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i!="",j=~"XXX|YYY"-10                3.74ms ± 0%    0.00ms ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~"X|Y|Z",j="foo"-10                2.04µs ± 0%    2.13µs ± 1%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",j="foo"-10                    172ms ± 1%     168ms ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~"1.+",j="foo"-10                  27.3ms ± 0%    26.5ms ± 1%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!="2",j="foo"-10             182ms ± 0%     179ms ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!~"2.*",j="foo"-10           199ms ± 0%     193ms ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="X",i=~"^.+$",i!~"^.*2.*$",j="foo"-10    1.09µs ± 0%    1.49µs ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"0xxx|1xxx|2xxx"-10                     55.1µs ± 0%    55.8µs ± 1%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"(0|1|2)xxx"-10                         55.2µs ± 1%    55.0µs ± 1%   ~     (p=1.000 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"[0-2]xxx"-10                           55.6µs ± 0%    55.2µs ± 1%   ~     (p=0.400 n=3+3)
BucketIndexReader_ExpandedPostings/i!~[0-2]xxx-10                             96.4ms ± 0%    96.4ms ± 0%   ~     (p=1.000 n=3+3)
BucketIndexReader_ExpandedPostings/i=~".*",_i!~[0-2]xxx-10                     100ms ± 1%      99ms ± 1%   ~     (p=0.400 n=3+3)
BucketIndexReader_ExpandedPostings/p!=""-10                                   20.1ms ± 1%    19.9ms ± 0%   ~     (p=0.200 n=3+3)

name                                                                        old alloc/op   new alloc/op   delta
BucketIndexReader_ExpandedPostings/n="X"-10                                     889B ± 0%      921B ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="X",j="foo"-10                             953B ± 0%     1065B ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="X",j!="foo"-10                            953B ± 0%     1065B ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/j=~"XXX|YYY"-10                              936B ± 0%      936B ± 0%   ~     (all equal)
BucketIndexReader_ExpandedPostings/j=~"X.+"-10                                1.87kB ± 0%    1.92kB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"X|Y|Z"-10                                984B ± 0%      984B ± 0%   ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1"-10                                   12.7MB ± 0%    12.7MB ± 0%   ~     (p=0.200 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",j="foo"-10                           19.8MB ± 0%    19.8MB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/j="foo",n="1"-10                           19.8MB ± 0%    19.8MB ± 0%   ~     (p=0.800 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",j!="foo"-10                          19.8MB ± 0%    19.8MB ± 0%   ~     (p=0.700 n=3+3)
BucketIndexReader_ExpandedPostings/i=~".*"-10                                  287MB ± 0%     287MB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~".+"-10                                  327MB ± 0%     327MB ± 0%   ~     (p=0.400 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"^.+$",j=~"X.+"-10                      8.01MB ± 0%    4.81MB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~""-10                                    113MB ± 0%     113MB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i!=""-10                                    327MB ± 0%     327MB ± 0%   ~     (p=0.200 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",j="foo"-10                   21.4MB ± 0%    21.4MB ± 0%   ~     (p=0.700 n=3+3)
BucketIndexReader_ExpandedPostings/n="X",i=~"^.+$",j="foo"-10                 1.00kB ± 0%    1.26kB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",i!="2",j="foo"-10            22.7MB ± 0%    22.7MB ± 0%   ~     (p=0.700 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i!=""-10                             86.4MB ± 0%    86.4MB ± 0%   ~     (p=0.200 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i!="",j="foo"-10                     93.5MB ± 0%    93.5MB ± 0%   ~     (p=1.000 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i!="",j=~"X.+"-10                    8.01MB ± 0%    4.81MB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i!="",j=~"XXX|YYY"-10                8.01MB ± 0%    0.00MB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~"X|Y|Z",j="foo"-10                1.30kB ± 0%    1.34kB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",j="foo"-10                   93.5MB ± 0%    93.5MB ± 0%   ~     (p=0.400 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~"1.+",j="foo"-10                  23.9MB ± 0%    23.9MB ± 0%   ~     (p=1.000 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!="2",j="foo"-10            94.8MB ± 0%    94.8MB ± 0%   ~     (p=0.300 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!~"2.*",j="foo"-10           103MB ± 0%     100MB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="X",i=~"^.+$",i!~"^.*2.*$",j="foo"-10    1.06kB ± 0%    1.47kB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"0xxx|1xxx|2xxx"-10                     14.1kB ± 0%    14.1kB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"(0|1|2)xxx"-10                         14.0kB ± 0%    14.0kB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"[0-2]xxx"-10                           14.0kB ± 0%    14.0kB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i!~[0-2]xxx-10                              285MB ± 0%     285MB ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~".*",_i!~[0-2]xxx-10                     287MB ± 0%     287MB ± 0%   ~     (p=0.200 n=3+3)
BucketIndexReader_ExpandedPostings/p!=""-10                                   58.8MB ± 0%    58.8MB ± 0%   ~     (p=0.700 n=3+3)

name                                                                        old allocs/op  new allocs/op  delta
BucketIndexReader_ExpandedPostings/n="X"-10                                     20.0 ± 0%      21.0 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="X",j="foo"-10                             20.0 ± 0%      22.0 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="X",j!="foo"-10                            20.0 ± 0%      22.0 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/j=~"XXX|YYY"-10                              22.0 ± 0%      22.0 ± 0%   ~     (all equal)
BucketIndexReader_ExpandedPostings/j=~"X.+"-10                                  21.0 ± 0%      22.0 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"X|Y|Z"-10                                23.0 ± 0%      23.0 ± 0%   ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1"-10                                     92.0 ± 0%      91.0 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",j="foo"-10                              122 ± 1%       117 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/j="foo",n="1"-10                              121 ± 0%       117 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",j!="foo"-10                             119 ± 1%       115 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~".*"-10                                    108 ± 0%       107 ± 1%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~".+"-10                                   600k ± 0%      600k ± 0%   ~     (p=1.000 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"^.+$",j=~"X.+"-10                        27.0 ± 0%      26.0 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~""-10                                     600k ± 0%      600k ± 0%   ~     (p=0.400 n=3+3)
BucketIndexReader_ExpandedPostings/i!=""-10                                     600k ± 0%      600k ± 0%   ~     (p=1.000 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",j="foo"-10                      126 ± 0%       122 ± 1%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="X",i=~"^.+$",j="foo"-10                   20.0 ± 0%      25.0 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",i!="2",j="foo"-10               154 ± 0%       147 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i!=""-10                               600k ± 0%      600k ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i!="",j="foo"-10                       600k ± 0%      600k ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i!="",j=~"X.+"-10                      33.0 ± 0%      29.0 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i!="",j=~"XXX|YYY"-10                  34.0 ± 0%      27.0 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~"X|Y|Z",j="foo"-10                  29.0 ± 0%      27.0 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",j="foo"-10                     600k ± 0%      600k ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~"1.+",j="foo"-10                   66.8k ± 0%     66.8k ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!="2",j="foo"-10              600k ± 0%      600k ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!~"2.*",j="foo"-10            667k ± 0%      667k ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/n="X",i=~"^.+$",i!~"^.*2.*$",j="foo"-10      20.0 ± 0%      27.0 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"0xxx|1xxx|2xxx"-10                        117 ± 0%       116 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"(0|1|2)xxx"-10                            117 ± 0%       116 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~"[0-2]xxx"-10                              117 ± 0%       116 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i!~[0-2]xxx-10                                167 ± 0%       164 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/i=~".*",_i!~[0-2]xxx-10                       172 ± 0%       169 ± 0%   ~     (p=0.100 n=3+3)
BucketIndexReader_ExpandedPostings/p!=""-10                                     98.0 ± 0%      98.0 ± 0%   ~     (all equal)
```

</details>